Move 'not-has-class' to private folder
